pyproject.toml
[project]
name = "fathom-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "llama-cpp-python>=0.3.16",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlite-vec>=0.1.7",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = ["ruff>=0.15.7"]
